+++ CHANGELOG.md
@@ -14,5 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add sound effects on the tile entities that are sorted
 - auto-sort target after move
 - move content from source to target on close
+- only sort if source is within LCB
+- repeat check for multiple blocks around source (not just y+1)
 - update signed target with denied reason
 - update signed target with moved item count